requirements/setup.txt
# additional package for integrating pytest in setuptools

pytest-runner==5.3.0
